requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==2.0.3
